DAGs/libs/algorithm/poh_network/pyproject.toml
# D:\city_chain_project\DAGs\libs\algorithm\poh_network\pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "poh_network"
version         = "0.1.0"
description     = "PoH network broadcast/receive manager"
authors         = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.8"
dependencies = [
    "poh_storage>=0.1.0",
    "grpcio>=1.51.0",
    "grpcio-tools>=1.51.0",
    "aiohttp>=3.8.0",
    "protobuf==5.27.0",          # ← ここで安全版に固定
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-asyncio>=1.0",
]

[tool.setuptools.packages.find]
where   = ["."]
include = ["poh_network*"]

[project.scripts]
poh-network = "poh_network.app_network:main"

[project.urls]
"Source"        = "https://example.com/your-repo"
"Bug Tracker"   = "https://example.com/your-repo/issues"

[tool.pytest.ini_options]
testpaths = ["poh_network/tests"]
